Move `Requests` & `AttacksDetected` to stats pkg
